+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/ConfigurationModifier.java
@@ -1,7 +1,7 @@
 /*
  * Tencent is pleased to support the open source community by making spring-cloud-tencent available.
  *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company. All rights reserved.
+ * Copyright (C) 2021 Tencent. All rights reserved.
  *
  * Licensed under the BSD 3-Clause License (the "License");
  * you may not use this file except in compliance with the License.
@@ -114,6 +114,8 @@ public class ConfigurationModifier implements PolarisConfigurationConfigModifier
 		}
 
 		connectorConfig.setAddresses(configAddresses);
+		connectorConfig.setLbPolicy(polarisContextProperties.getAddressLbPolicy());
+		connectorConfig.setServerSwitchInterval(polarisContextProperties.getServerSwitchInterval());
 
 		if (StringUtils.isNotEmpty(polarisConfigProperties.getToken())) {
 			connectorConfig.setToken(polarisConfigProperties.getToken());