+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/config/PolarisContextAutoConfiguration.java
@@ -1,7 +1,7 @@
 /*
  * Tencent is pleased to support the open source community by making spring-cloud-tencent available.
  *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company. All rights reserved.
+ * Copyright (C) 2021 Tencent. All rights reserved.
  *
  * Licensed under the BSD 3-Clause License (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,8 +46,14 @@ import org.springframework.core.env.Environment;
  * @author Haotian Zhang
  */
 @ConditionalOnPolarisEnabled
-@EnableConfigurationProperties({PolarisContextProperties.class})
+@EnableConfigurationProperties
 public class PolarisContextAutoConfiguration {
+
+	@Bean
+	@ConditionalOnMissingBean
+	public PolarisContextProperties polarisContextProperties() {
+		return new PolarisContextProperties();
+	}
 
 	@Bean(initMethod = "init")
 	@ConditionalOnMissingBean